requirements.txt
flask
python-dotenv
numpy
pandas
geopandas
scikit-learn==1.6.1